fixed warnings when compiling for 64 bit architecture removed PureData preset

+++ TangoHapps/classes/objects/software/THContactBook.m
@@ -282,7 +282,7 @@ NSString * const kCallImageName = @"call.png";
     CFArrayRef people = ABAddressBookCopyArrayOfAllPeople(_addressBook);
     _contacts = CFArrayCreateMutableCopy(kCFAllocatorDefault,CFArrayGetCount(people), people);
     
-    CFArraySortValues(_contacts, CFRangeMake(0, CFArrayGetCount(_contacts)), (CFComparatorFunction)ABPersonComparePeopleByName, (void*) ABPersonGetSortOrdering());
+    CFArraySortValues(_contacts, CFRangeMake(0, CFArrayGetCount(_contacts)), (CFComparatorFunction)ABPersonComparePeopleByName, (void*) (NSInteger) ABPersonGetSortOrdering());
     
     CFRelease(people);
